refactor: add pushNotification for step 4
